+++ src/page/EditExam.js
@@ -1,7 +1,9 @@
 // import { useRef } from "react";
 import EditorExam from "../components/EditorExam"
 import AttributeManager from "../components/AttributeManager"
-import { useState } from "react";
+import AICreate from "../test/AI/AICreate";
+import { useState, useEffect } from "react";
+import { useLocation } from "react-router-dom";
 import { useLoginController } from "../function/useLoginController";
 // import { useDataHandle } from "../..//dataHandle";
 
@@ -11,6 +13,29 @@ export default function EditExam() {
   const { handleAutoLogin, handleLogout, handleLoginState } = useLoginController();
 
   const [isCreate, setCreate] = useState([]);
+
+  const location = useLocation();
+
+  // 선택된 시험의 examId와 examTitle을 받아, AttributeManager 컴포넌트에게 props로 전달
+  const [examTitle, setExamTitle] = useState('');
+  const [examId, setExamId] = useState('');
+
+  console.log("Location:", location);
+  console.log("Exam ID:", examId);
+  console.log("Exam Title:", examTitle);
+
+  useEffect(() => {
+    // location.state가 정의되어 있을 때만 examId와 examTitle을 설정합니다.
+    if (location && location.state) {
+      const { examId, examTitle } = location.state;
+      setExamId(examId);
+      setExamTitle(examTitle);
+      console.log("Location:", location);
+      console.log("Exam ID:", examId);
+      console.log("Exam Title:", examTitle);
+
+    }
+  }, [location]);
 
   const handleCreate = () => {
     // 새로운 컴포넌트를 생성하여 상태에 추가
@@ -27,14 +52,15 @@ export default function EditExam() {
   return (
     <>
       <h1>Test</h1>
-      <div style={{ marginBottom: '40px' }}>
+      {/* <div style={{ marginBottom: '40px' }}>
         <button style={{ backgroundColor: 'gray', color: 'white' }} onClick={() => { handleAutoLogin() }}>logIn</button>
         <button style={{ backgroundColor: 'gray', color: 'white' }} onClick={() => { handleLogout() }}>logOut</button>
         <button style={{ backgroundColor: 'gray', color: 'white' }} onClick={() => { handleLoginState() }}>logState</button>
-      </div>
-      <button>시험지 생성</button>
-      <AttributeManager></AttributeManager>
+      </div> */}
+      
+      <AttributeManager examId={examId} />
       {/* AttributeManager 저장하기가 되어야 EditorExam이 활성화 */}
+      <br /> <br />
 
       <button
         onClick={() => { handleCreate(); console.log(isCreate.length); }}
@@ -49,6 +75,7 @@ export default function EditExam() {
           </div>
         </div>
       ))}
+      <AICreate examId={examId}/>
 
 
     </>
